requirements.txt
asgiref==3.6.0
Django==4.1.5
Pillow==8.4.0
sqlparse==0.4.3
termcolor==2.2.0
